Implement authentication context and provider
- Refactored authentication handling by creating AuthContext and AuthProvider components.
- Updated App component to use the new AuthProvider.
- Enhanced useAuth hook to ensure proper context usage.
- Added Firebase authentication methods for sign-in and sign-up, including OAuth support for Google, Facebook, and Apple.
- Improved error handling and loading states in authentication processes.
- Updated Firebase config validation to accommodate testing environments.
