+++ src/contexts/AuthContext.ts
@@ -1,0 +1,5 @@
+import { AuthContextType } from "@/lib/firebase/types";
+import { createContext } from "react";
+
+// Create the context
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
