+++ src/hooks/useAuth.ts
@@ -1,2 +1,11 @@
-// Re-export the useAuth hook from the context for convenience
-export { useAuth } from '@/contexts/AuthContext';
+import { AuthContextType } from "@/lib/firebase/types";
+import { AuthContext } from "@/contexts/AuthContext";
+import { useContext } from "react";
+
+export const useAuth = (): AuthContextType => {
+  const context = useContext(AuthContext);
+  if (context === undefined) {
+    throw new Error('useAuth must be used within an AuthProvider');
+  }
+  return context;
+};
